requirements.txt
aiogram==2.25.1
# Примечание: aiohttp 3.13.2 используется для совместимости с Python 3.13
# aiogram 2.25.1 требует aiohttp<3.9.0, но это не работает на Python 3.13
# Версия 3.13.2 работает с aiogram 2.25.1 на практике
aiohttp>=3.13.0
fastapi==0.104.1
uvicorn==0.24.0
# SQLAlchemy 2.0.44 требуется для совместимости с Python 3.13
sqlalchemy>=2.0.44
aiosqlite==0.19.0
python-dotenv==1.0.0
telethon==1.34.0
pytz>=2024.1
requests>=2.31.0
